requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
hupper==1.12.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keyboard==0.13.5
MarkupSafe==2.1.3
MouseInfo==0.1.3
numpy==1.26.3
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==15.0.0
PyAutoGUI==0.9.54
PyDirectInput==1.0.4
PyGetWindow==0.0.9
pymongo==4.6.1
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2024.1
reportlab==4.0.9
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
stripe==8.2.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
waitress==3.0.0
Werkzeug==3.0.1
whitenoise==6.6.0
